recover TODO items and rename catalog trait
