+++ src/moonlink/src/storage/iceberg.rs
@@ -1,8 +1,8 @@
 pub(crate) mod catalog_utils;
 pub(crate) mod deletion_vector;
-pub(crate) mod deletion_vector_trait;
 pub(crate) mod file_catalog;
 pub(crate) mod iceberg_table_manager;
+pub(crate) mod moonlink_catalog;
 pub(crate) mod object_storage_catalog;
 pub(crate) mod puffin_utils;
 pub(crate) mod puffin_writer_proxy;
